feat: name response object by resource class
refs: https://github.com/RonasIT/laravel-swagger/issues/192
