+++ src/Support/Resolvers/ResourceClassResolver.php
@@ -2,15 +2,41 @@
 
 namespace RonasIT\AutoDoc\Support\Resolvers;
 
+use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Http\Resources\Json\JsonResource;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Str;
 use ReflectionFunctionAbstract;
+use ReflectionNamedType;
+use ReflectionUnionType;
 use RonasIT\AutoDoc\DTO\ResolvedResource;
 
 class ResourceClassResolver
 {
     public function resolve(ReflectionFunctionAbstract $reflection): ?ResolvedResource
+    {
+        return $this->resolveFromReturnType($reflection->getReturnType())
+            ?? $this->resolveFromSource($reflection);
+    }
+
+    private function resolveFromReturnType(mixed $returnType): ?ResolvedResource
+    {
+        if ($returnType instanceof ReflectionNamedType && $this->isResourceClass($returnType->getName())) {
+            return new ResolvedResource($returnType->getName());
+        }
+
+        if ($returnType instanceof ReflectionUnionType) {
+            foreach ($returnType->getTypes() as $type) {
+                if ($type instanceof ReflectionNamedType && !$type->isBuiltin() && $this->isResourceClass($type->getName())) {
+                    return new ResolvedResource($type->getName());
+                }
+            }
+        }
+
+        return null;
+    }
+
+    private function resolveFromSource(ReflectionFunctionAbstract $reflection): ?ResolvedResource
     {
         $fileContent = $this->getFileContent($reflection);
         $code = $this->getFunctionCode($reflection, $fileContent);
@@ -28,11 +54,9 @@ class ResourceClassResolver
                 continue;
             }
 
-            $resourceName = $matches[1];
-
-            $resourceName = class_exists($resourceName)
-                ? $resourceName
-                : $this->getClassNameFromImports($resourceName, $fileContent);
+            $resourceName = class_exists($matches[1])
+                ? $matches[1]
+                : $this->getClassNameFromImports($matches[1], $fileContent);
 
             if (is_subclass_of($resourceName, JsonResource::class)) {
                 return new ResolvedResource($resourceName, $type === 'collection');
@@ -40,6 +64,12 @@ class ResourceClassResolver
         }
 
         return null;
+    }
+
+    private function isResourceClass(string $className): bool
+    {
+        return is_subclass_of($className, JsonResource::class)
+            && $className !== AnonymousResourceCollection::class;
     }
 
     private function getFileContent(ReflectionFunctionAbstract $reflection): array
